% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encrypt.data.R
\name{encrypt.data}
\alias{encrypt.data}
\title{Encrypt data from \code{get.data} or \code{is.get.fata.output}}
\usage{
encrypt.data(data, germplasm_name = "germplasm", person_name = "person")
}
\arguments{
\item{data}{output from \code{get.data}}

\item{germplasm_name}{Name of the germplasms in the encrypted data. "germplasms" by default}

\item{person_name}{Name of the persons an selection persons in the encrypted data. "persons" by default}
}
\value{
The function returns the data frame encrypted and writes in the work directory a file.RData with the key
}
\description{
\code{encrypt.data} encrypts data from \code{get.data} or \code{is.get.fata.output}
}
\details{
The following information is encrypted:
person,
year (from 2000),
project,
selection_person,
reproduction_type,
germplasm_type,
germplasm and
seed-lots

All query.type can be encrypt unless "species", "variable", "reproduction.type" and "methods"

The key is written in .RData format. To use it you should type load("key.RData").
}
\examples{
# See the vignette
}
\author{
Pierre Rivière
}
\seealso{
\code{\link{get.data}}, \code{\link{is.get.data.output}}
}

