% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{extra_functions}
\alias{data.to.use}
\alias{extra_functions}
\title{Some functions used in several functions of shinemas2R}
\usage{
data.to.use(data, correlated_group)
}
\description{
This file group several functions used in several functions of shinemas2R:
\itemize{
	\item get.ggplot
	\item get.table
}
}
\author{
Pierre Riviere
}

