% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.data.R
\name{get.data}
\alias{get.data}
\title{Query SHiNeMaS and return the datasets}
\usage{
get.data(db_user = "pierre", db_host = "127.0.0.1", db_name = "shinemas",
  db_password = "pierre", query.type = "person",
  fill.diffusion.gap = FALSE, network.info = TRUE, Mdist = FALSE,
  data.type = NULL, filter.on = NULL, germplasm.in = NULL,
  germplasm.out = NULL, germplasm.type.in = NULL,
  germplasm.type.out = NULL, year.in = NULL, year.out = NULL,
  project.in = NULL, project.out = NULL, person.in = NULL,
  person.out = NULL, seed.lot.in = NULL, seed.lot.out = NULL,
  relation.in = NULL, reproduction.type.in = NULL, variable.in = NULL)
}
\arguments{
\item{db_user}{user name of SHiNeMaS}

\item{db_host}{IP adress of the computer where SHiNeMaS is!, If local host db_host = "127.0.0.1"}

\item{db_name}{name fo the data base}

\item{db_password}{your password to login. If no password is needed, put ""}

\item{query.type}{Type of query, which will create a data set. There are five types:
\enumerate{

\item "network": network relations between seed-lots

\item "SL.mix": seed-lots merged from replications and not from 'real' mixture

\item "cross": seed-lots used to give a cross: father, grandfather, mother, grandmother and cross.

\item raw information on levels and variables present in SHiNeMaS :
\itemize{
\item "variable"
\item "person"
\item "year"
\item "project"
\item "seed.lot"
\item "selection.person": persons that did intra-varietal mass selection
\item "reproduction.type"
\item "germplasm.type"
\item "germplasm"
}

\item data with variable on specific seed-lots :
\itemize{
\item "data-classic": default dataframe
\item "data-S": selection differential for a given year.
\item "data-SR": seed-lots pairs of selection differential for year n-1 and response to selection for year n.
}

}}

\item{fill.diffusion.gap}{For query.type = "network", create a network with no gaps between seed-lots (as long as there is information!)}

\item{network.info}{For query.type = "network", if TRUE, aggregate information on relation and seed-lots on the network.}

\item{Mdist}{For query.type = "network", if TRUE, compute the Mdist matrix. See details.}

\item{data.type}{For queries in "data-", type of data: "relation" for data linked to relation between seed lots and "seed-lots" for data linked to seed lots}

\item{filter.on}{This argument is needed for filters for queries in "network", data-" and "data.type" = data.relation. It chooses on which seed-lots the filters is applied: "son", "father" or "father-son".}

\item{germplasm.in}{Filter: vector with germplasms to keep. By default, all the germplasms are in.}

\item{germplasm.out}{Filter: vector with germplasms to discard. By default, no germplasms are out.}

\item{germplasm.type.in}{Filter: vector with germplasms to keep. By default, all the germplasm types are in.}

\item{germplasm.type.out}{Filter: vector with germplasms to discard. By default, no germplasm types are out.}

\item{year.in}{Filter: vector with years to keep. By default, all the years are in.}

\item{year.out}{Filter: vector with years to discard. By default, no years are out.}

\item{project.in}{Filter: vector of projects to keep. By default, all the projects are in.}

\item{project.out}{Filter: vector of projects to discard. By default, no projects are out.}

\item{person.in}{Filter: vector of persons to keep. By default, all the persons are in.}

\item{person.out}{Filter: vector of persons to discard.  By default, no persons are out.}

\item{seed.lot.in}{Filter: vector of seed-lots to keep. By default, all the seed-lots are in.}

\item{seed.lot.out}{Filter: vector of seed-lots to discard. By default, on seed-lots are out.}

\item{relation.in}{Filter on a relation to keep: "reproduction", "mixture", "selection" or "diffusion". By default, all relations are in.}

\item{reproduction.type.in}{Filter on reproduction types}

\item{variable.in}{Filter: vector with variables to keep. If NULL, all variables in SHiNeMaS are displayed.}
}
\value{
The function returns a list with

\itemize{
\item data according to the query type.
\enumerate{

\item For network it returns a list with
\itemize{
\item the netwok object
\item the network.info matrix with information on relation in the network
\item the Mdist square matrix with the number of reproductions that separate two seed-lots since their last common diffusion.
}

\item For SL.mix, it returns a data frame with seed-lots son (i.e. coming from the mixture), seed-lots father, germplasm, person, year and project

\item For raw information on levels and variables present in SHiNeMaS it returns a vector

\item For data with variable on specific seed-lots it returns a list with
\itemize{
		\item the data set with correlated variables
		\item the data set with non correlated variables
		\item the description of the methodes used for each variable
		}

}

\item information to connect to SHiNeMaS
}
}
\description{
\code{get.data} queries SHiNeMaS and return the data-sets.
}
\details{
\itemize{

\item Queries "cross"

The data frame returned has a column "expe" which corresponds to an id of the set of seed-lots used in the cross


\item Queries "S"

The data frame returned has a column "expe" which corresponds to an id of one selection differential

\item Queries "SR"

The data frame returned has a column "expe" which corresponds to an id of one selection differential and the corresponding response to selection

The query "data-SR" takes into account when selection have been done in a seed lot, that this seed lot have been merged and then have been sown. It is the case when selection have been carried out in a replication that have been merge after. Even if this case should not arrise, it may happen.


\item Query "network"

The Mdist square matrix can be compared to a differenciation distance. It can be put in relation with genetic Fst for example.
}
}
\author{
Pierre Riviere for R code and Yannick de Oliveira for SQL code
}
\seealso{
\code{\link{encrypt.data}}, \code{\link{translate.data}},, \code{\link{get.ggplot}}, \code{\link{get.table}}
}

