% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot.R
\name{get.ggplot}
\alias{get.ggplot}
\title{Get ggplot objects from \code{get.data}}
\usage{
get.ggplot(data, correlated_group = NULL, fuse_g_and_s = FALSE,
  ggplot.type = NULL, ggplot.display = NULL, ggplot.on = "son",
  x.axis = NULL, in.col = NULL, vec_variables = NULL,
  nb_parameters_per_plot_x.axis = NULL,
  nb_parameters_per_plot_in.col = NULL, vertex.size = 3,
  vertex.color = "year", organise.sl = FALSE, labels.on = "son",
  hide.labels.parts = "all", display.labels.sex = TRUE,
  labels.generation = TRUE, labels.size = 3, location.map = "france",
  pie.size = 0.5)
}
\arguments{
\item{data}{output from get.data with query.type = "network" or query.type = "data-...".}

\item{correlated_group}{Name of the group of correlation in data. NULL by default.}

\item{fuse_g_and_s}{Fuse germplasm and selection name information in a column named germplasm}

\item{ggplot.type}{the type of plot you wish.
NULL by default: according to the type of data, it will be "network-all" or "data-all". "X-all" means that all the "X-" objects are done.
The possibilites are:
\itemize{
\item for "network-":
"network-network",
"network-reproduction-harvested",
"network-reproduction-positive-inter-selected",
"network-reproduction-negative-inter-selected",
"network-reproduction-crossed",
"network-diffusion-sent",
"network-diffusion-received",
"network-diffusion-relation",
"network-mixture",
"network-positive-intra-selected",
"network-Mdist".

\item for "data-":
\itemize{
        \item "data-barplot"
        \item "data-boxplot"
        \item "data-interaction"
        \item "data-radar"
        \item "data-biplot"
        \item "data-pie.on.map" (only for data-classic).
        \item "data-pie.on.network"    (only for data-classic).
        }
}}

\item{ggplot.display}{For "network-" data: "barplot", "map". It can be a vector of several elements i.e. c("barplot", "map"). NULL by default: both are done.}

\item{ggplot.on}{For "data-" type. father" or "son" depending on which seed-lot you want to plot.}

\item{x.axis}{factor display on the x.axis of a plot: "germplasm", "year" or "person" referring to the attributes of a seed-lots. If NULL, all the combinaison are done for x.axis and in.col.}

\item{in.col}{factor display in color of a plot: "germplasm", "year" or "person" referring to the attributes of a seed-lots. If NULL, in.col is not displayed. Note it is compulsory for data-biplot and data-radar as in these cases x.axis is not used.}

\item{vec_variables}{For "data-" type : a vector of variables displayed.}

\item{nb_parameters_per_plot_x.axis}{The number of parameters per plot on x axis arguments}

\item{nb_parameters_per_plot_in.col}{The number of parameters per plot for in.col arguments}

\item{vertex.size}{For ggplot.type = "network", size of the vertex}

\item{vertex.color}{For ggplot.type = "network", color of the vertex. It can be chosen according to  "person", "germplasm" or "year". If NULL, it is in black.}

\item{organise.sl}{For ggplot.type = "network", organise seed-lots for an easier visualisation}

\item{labels.on}{"son" or "father". For ggplot.type = "data-biplot" choose if labels are on son or father seed-lots. For other "data-..." plots, there are no labels as it is managed by x.axis and in.col.}

\item{hide.labels.parts}{For ggplot.type = network-network and data-biplot.
Parts of the label hidden: "germplasm", "person", "year", "person:germplasm", "year:germplasm", "person:year", "all". "all" means that no labels are dispayed. If NULL labels are displayed. Labels are based on seed-lots names under the form germplasm_year_person_digit.
For "data-biplot", the default value is NULL.
For easier visualisation, Digit is never display unless you choose NULL.
Note that with ggplot.type "network-" and ggplot.display = "map", hide.labels.part can be NULL or "all" as only person can be display.}

\item{display.labels.sex}{For ggplot.type = "network-network", if TRUE, display the sex of the seed-lot if it has been used in a cross. Nothing is displayed if hide.labels.parts = "all".}

\item{labels.generation}{For ggplot.type = "network-network", if TRUE, display generation for each reproduction}

\item{labels.size}{Size of the labels}

\item{location.map}{Location of the map see \code{?map} for more details.}

\item{pie.size}{Size of the pie when using pies}
}
\value{
The function returns a list with several ggplot objects according to the data and ggplot.type used.

\enumerate{

\item For "network-" type it returns a list with ggplot objects representing the network and for each relation (i) barplots with the number of populations (ii) maps for each year

	\itemize{
	\item the network (ggplot.type = "network-network")

	\item a series of ggplots for reproduction:
		\itemize{
		\item harvested seed-lots (ggplot.type = "network-reproduction-harvested")
		\item positive inter selected seed-lots (ggplot.type = "network-reproduction-positive-inter-selected")
		\item negative inter selected seed-lots (ggplot.type = "network-reproduction-negative-inter-selected")
		\item seed-lots involved in crosses (ggplot.type = "network-reproduction-crossed")
		}

\item a series of ggplots for diffusion representing:
	\itemize{
	\item the seed-lots sent (ggplot.type = "network-diffusion-sent")
	\item the seed-lots received (ggplot.type = "network-diffusion-received")
	\item the relation between seed-lots on maps (ggplot.type = "network-diffusion-relation"). The size of the arrows is proprotionnal to the number of diffusions.
	}

\item a series of ggplots for mixture (ggplot.type = "network-mixture"). Note that mixture of replications are not counted here.

\item a series of ggplots for selection (ggplot.type = "network-positive-intra-selected")

\item, If Mdist is not NULL, draw ??? Cf ma thèse !!!0
}

\item For "data-" type, plots are displayed for each variable of vec_variables for "data-barplot", "data-boxplot", "data-interation". It is all possibles pairs for "data-biplot". It is all the variables for "data-radar".

}
}
\description{
\code{get.ggplot} generates ggplot object from \code{get.data} data-set
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{get.data}}
}

