% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_hide.labels.part.R
\name{get.ggplot_hide.labels.part}
\alias{get.ggplot_hide.labels.part}
\title{Function use in get.ggplot}
\usage{
get.ggplot_hide.labels.part(label, hide.labels.parts)
}
\description{
hide parts of seed-lots labels
}

