% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_network.relation.barplot.R
\name{get.ggplot_network.relation.barplot}
\alias{get.ggplot_network.relation.barplot}
\title{Function used in get.ggplot}
\usage{
get.ggplot_network.relation.barplot(data, combi, relation = "reproduction",
  relation.type = "recolte", name = "harvested seed-lots",
  nb_parameters_per_plot_x.axis, nb_parameters_per_plot_in.col, ggplot.type)
}
\description{
draw barplot for network relations
}

