% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_network.relation.map.R
\name{get.ggplot_network.relation.map}
\alias{get.ggplot_network.relation.map}
\title{Function used in get.ggplot}
\usage{
get.ggplot_network.relation.map(map, data, relation = "reproduction",
  relation.type = "recolte", name = "harvested seed-lots", pie.size,
  hide.labels.parts, labels.size)
}
\description{
draw pie on maps for network relations
}

