% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_plot.it.r
\name{get.ggplot_plot.it}
\alias{get.ggplot_plot.it}
\title{Function use in get.ggplot}
\usage{
get.ggplot_plot.it(d, titre = NULL, ggplot.display, x.axis, in.col,
  ggplot.type, vec_variables, equal.ylim = TRUE,
  nb_parameters_per_plot_x.axis = NULL,
  nb_parameters_per_plot_in.col = NULL, labels.on = NULL,
  hide.labels.parts = NULL, labels.size = 1)
}
\description{
get ggplots: barplot, boxplots, interaction plots, radar, biplot
}

