% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_plot.network.R
\name{get.ggplot_plot.network}
\alias{get.ggplot_plot.network}
\title{Function use in get.ggplot}
\usage{
get.ggplot_plot.network(data, vertex.color, vertex.size, hide.labels.parts,
  display.labels.sex, labels.generation, organise.sl, labels.size = 1)
}
\description{
get network representation in ggplot
}

