% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.pdf.R
\name{get.pdf}
\alias{get.pdf}
\title{Get a pdf by concatene outputs from \code{get.ggplot} and \code{get.table}}
\usage{
get.pdf(dir = "./", form.name, LaTeX_head = NULL, LaTeX_body,
  compile.tex = TRUE, color1 = "white", color2 = "white")
}
\arguments{
\item{dir}{directory where the pdf should be created}

\item{form.name}{Name of the pdf file and the folder that contains all the raw information. See details.}

\item{LaTeX_head}{Path of tex file with head of the LaTeX document. It contains the packages and macros used in LaTeX_body. If NULL, a default LaTeX head is generated.}

\item{LaTeX_body}{A list that contains the elements to be displayed in the pdf.
This list must contains the following elements:

\itemize{
    \item "titlepage" : a list containing the following elements :
    \itemize{
        \item "title" : the title of the document. Nothing by default.
        \item "authors" : the authors of the document. Nothing by default.
        \item "email" : the corresponding email. Nothing by default.
        \item "date" : the date of the document. Nothing by default.
    }

    \item "tableofcontents" : if TRUE, a table of content is display.

    \item "chapter" : name of a chapter

    \item "section" : name of a section

    \item "subsection" : name of a subsection

    \item "subsubsection" : name of a subsubsection

    \item "text" : text to add

    \item "includepdf" : path of a .pdf to include

    \item "input" : path of a .tex to insert

    \item "table" : a list containing the following elements :
    \itemize{
        \item "content" : output from \code{get.table}
        \item "caption" : caption of the table
        \item "landscape" : If TRUE, the table is in landscape. FALSE by default.
        \item "display.rownames" : If TRUE, display the row names of the table. FALSE by default.
    }

    \item "figure" : a list containing the following elements :
    \itemize{
        \item "content" : output from \code{get.ggplot}
        \item "caption" : caption of the figure
        \item "layout" : the layout of the plots. It is a matrix under the form layout = matrix(c(1:4), ncol = 2, nrow = 2). It is layout = matrix(1, ncol = 1, nrow= 1) by default.
        \item "width" : width of the figure in textwidth unit. 1 means as width as the text. It is 1 by default.
        \item "landscape" : If TRUE, the figure is in landscape. FALSE by default.
    }
    }}

\item{compile.tex}{If TRUE compile the pdf.}

\item{color1}{For table, color of the head of the table. White by default.}

\item{color2}{For table, color of the row. White by default. The color of the rows is an alternation of white and color2.}
}
\value{
The function returns a pdf which is the result of the concatenation as well as folders with raw data (see Details).
}
\description{
\code{get.pdf} generates a pdf file by using LaTeX
}
\details{
The function creates one folder named "form.name" (first argument of the function).
This folder contains two folders: "figures" where all the pdf of the plots are stored and "tex.files" where all the .tex generated are stored.
The pdf generated is moved from "./form.name/tex.files" to the current directory.

Note this function works if you have installed LaTeX and the following packages: longtable, lscape, graphicx, pdfpages, float, hyperref and fancyhdr.
To download LaTeX, go to http://latex-project.org/ftp.html

Note that the argument [H] from the package float is used by default for figures and tables.

If you use LaTeX macro, do not forget to use escape mode, for example \code{\\textbf{}}.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{get.ggplot}}, \code{\link{get.table}}
}

