% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.table.R
\name{get.table}
\alias{get.table}
\title{Get table from \code{get.data}}
\usage{
get.table(data, correlated.variables = TRUE, fuse_g_and_s = FALSE,
  table.type = NULL, table.on = "son", vec_variables = NULL,
  nb_row = NULL, nb_col = NULL, nb_duplicated_rows = 10,
  col_to_display = c("person", "germplasm", "year", "block", "X", "Y"),
  invert_row_col = FALSE)
}
\arguments{
\item{data}{It can be either output from \code{get.data} or ggplot.object from \code{get.ggplot}.}

\item{correlated.variables}{For data from \code{get.data}. If TRUE, get datasets.with.correlated.variables. If FALSE, get datasets.with.non.correlated.variables}

\item{fuse_g_and_s}{Fuse germplasm and selection name information in a column named germplasm}

\item{table.type}{the type of table you wish according to the type of data.
For data coming from \code{get.data}, it can be :
    \itemize{
    \item "raw" display raw data. Useful with text for example
    \item "mean" display for each variable columns with mean
    \item "mean.sd" display for each variable columns with mean and standard deviation
    \item "mean.sd.cv" display for each variable columns with mean, standard deviation and coefficient of variation
    \item "summary" display "Min.", "1st Qu.", "Median", "3rd Qu.", "Max." of the data
    }}

\item{table.on}{For "data-" type. father" or "son" depending on which seed-lot you want to display.}

\item{vec_variables}{Vector of variables displayed}

\item{nb_row}{The number of rows in the table}

\item{nb_col}{The number of columns in the table. col_to_display remains fixed.}

\item{nb_duplicated_rows}{Minimum number of duplicated rows for each variable of a table up to which the information is put in only one row.}

\item{col_to_display}{Columns to display in the table. It can be a vector with "person", "germplasm", "year", "block", "X" and "Y". If NULL, none of these columns are displayed. The variables follow these columns. For data-S and data-SR type, the column "expe" and "sl_statut" are added by default.}

\item{invert_row_col}{If TRUE, invert row and col in the table. This is possible only for col_to_display = NULL.}
}
\value{
The function returns a list with two elements
\itemize{
\item "duplicated_infos" : lists of two elements with seed-lots involved and variable values
\item "not_duplicated_infos": a list with the table of non duplicated information
}
}
\description{
\code{get.table} generates tables object from \code{get.data} data-set or ggplot.objet from \code{get.ggplot}.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{get.data}}, \code{\link{get.ggplot}}
}

