% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.get.data.output.R
\name{is.get.data.output}
\alias{is.get.data.output}
\title{Check if data set can be used as entry for \code{get.ggplot} and \code{get.table}}
\usage{
is.get.data.output(data, shinemas2R.object)
}
\arguments{
\item{data}{the date set to check}

\item{shinemas2R.object}{The type of shinemas2R.object to check. Possible values are:
\itemize{
    \item "network"
    \item "data-classic-seed-lots"
    \item "data-classic-relation"
    \item "data-S-seed-lots"
    \item "data-S-relation"
    \item "data-SR-seed-lots"
    \item "data-SR-relation"
    }}
}
\value{
The function returns a data set with the corresponding shinemas2R.object attributes or an error.
}
\description{
\code{is.get.data.output} check if data set can be used as entry for \code{get.ggplot} and \code{get.table}.
This is useful when the user do not have SHiNeMaS but wish to use the other functions of the packages: \code{get.ggplot}, \code{get.table} and \code{get.pdf}.
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{get.ggplot}}, \code{\link{get.table}}
}

