% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinemas2R-package.r
\docType{package}
\name{shinemas2R}
\alias{shinemas2R}
\alias{shinemas2R-package}
\title{shinemas2R}
\description{
The R package \code{shinemas2R} analyses outputs from the data base SHiNeMaS (Seeds History and Network Management System). It permits desciptive analysis as well as formatting data for existing R packages to perform statistical analysis.
}

