% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extdata.R
\name{shinemas_tuto.sql}
\alias{shinemas_tuto.sql}
\title{SHiNeMaS in sql format used in the tutorial of the vignette}
\format{sql}
\usage{
See the vignette Appendix A Install SHiNeMaS on localhost
}
\description{
shinemas_tuto.sql is full with data coming from the tutorial of SHiNeMaS
}
\author{
Pierre Rivière
}

