% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/translate.data.R
\name{translate.data}
\alias{translate.data}
\title{Translate variables and methods present in data coming from \code{get.data}}
\usage{
translate.data(data, list_translation)
}
\arguments{
\item{data}{output from the \code{get.data} function}

\item{list_translation}{A list with as many element as variables or methods to translate. It is under the form list(c("tkw", "pmg"), c("protein", "proteine")).}
}
\value{
The data with translated data
}
\description{
Translate variables and methods present in data coming from \code{get.data}
}
\seealso{
\code{\link{get.data}}, \code{\link{is.get.data.output}}
}

