% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.data.R
\name{format.data}
\alias{format.data}
\title{Format data from \code{get.data} to be used by several R packages}
\usage{
\method{format}{data}(data, data.on = "son", fuse_g_and_s = FALSE,
  correlated_group = NULL, format = "PPBstats")
}
\arguments{
\item{data}{output from the \code{get.data} function}

\item{data.on}{"son" or "father" depending on which seed-lots you want the information.}

\item{fuse_g_and_s}{Fuse germplasm and selection name information in a column named germplasm}

\item{correlated_group}{Name of the group of correlation in data. NULL by default meaning that all the data are taken.}

\item{format}{the format you want. It is under the form "name_of_the_package" or "name_of_the_package::name_of_the_function" if there are several formats within a package. Possible values are "PPBstats".}
}
\value{
The data set with the right format for the R package
}
\description{
Format data from \code{get.data} to be used by several R packages
}
\examples{
# See the vignette
}
\seealso{
\code{\link{get.data}}, \code{\link{is.get.data.output}}
}

