% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ggplot_pie.on.ggplot.R
\name{get.ggplot_pie.on.ggplot}
\alias{get.ggplot_pie.on.ggplot}
\title{Display pies on ggplot object such as network or map}
\usage{
get.ggplot_pie.on.ggplot(ggplot, data, variable, factor, x.origin = 1,
  y.origin = 1, r = 1, hide.labels.parts = NULL, labels.size = 2,
  precision = 10000)
}
\arguments{
\item{ggplot}{a ggplot object on wich put the pies}

\item{data}{}

\item{variable}{}

\item{factor}{}

\item{x.origin}{the x origin of the pie}

\item{y.origin}{the y origin of the pie}

\item{r}{the radius of the pie. If the radius is too big and the pie do not fit on the ggplot object, an error is returned.}

\item{hide.labels.parts}{}

\item{precision}{The precision in the drawing of the pie. The higher the better.}

\item{label.size}{label correspond to factor}
}
\value{
The function returns a ggplot with pies on it
}
\description{
\code{get.ggplot_pie.on.ggplot} display pies on ggplot object such as network or map
}
\examples{
to do
}
\author{
Pierre Rivière
}
\seealso{
\code{\link{get.ggplot}}
}

