% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.get.data.output.R
\name{is.get.data.output}
\alias{is.get.data.output}
\title{Check if data set can be used as entry for \code{get.ggplot} and \code{get.table}}
\usage{
is.get.data.output(data, shinemas2R.object)
}
\arguments{
\item{data}{the date set to check}

\item{shinemas2R.object}{The type of shinemas2R.object to check. Possible values are:
\itemize{
    \item "network"
    \item "data-classic-seed-lots"
    \item "data-classic-relation"
    \item "data-S-seed-lots"
    \item "data-S-relation"
    \item "data-SR-seed-lots"
    \item "data-SR-relation"
    }}
}
\value{
The function returns a data set with the corresponding shinemas2R.object attributes or an error.
}
\description{
\code{is.get.data.output} check if data set can be used as entry for \code{get.ggplot} and \code{get.table}.
This is useful when the user do not have SHiNeMaS but wish to use the other functions of the package.
}
\details{
Formats of the different shinemas2R.object:
	\itemize{
	 \item "network" It must be a list of three elements: "network", "network.info" and "Mdist" (see \code{get.data} for more details) with the followinf format:
	 \itemize{
	 	\item "network" must be a network object. It can be NULL. Vertex attributes can be put on "year", "person", "germplasm", germplasm.type" and "sex". Edge attributes can be put on "relation" and "generation"
	 	\item "network.info" must be a dataframe. It can be NULL. It has the following columns: sl, which is a character; alt, long and lat which are numeric; diffusion, id.diff, reproduction, mixture, cross.info, germplasm, person and year, which are factor.
	 	Possible values for diffusion are for EN: "give", "receive", "give-receive" or NA.
	 	Possible values for reproduction are for EN: "sow", "harvest", "harvest-sow" or NA.
	 	Possible values for selection are for EN: selection or NA.
	 	Possible values for mixture are for EN: mixture or NA.

	 	\item "Mdist" must be a square matrix with the same columns and rows names.
	 	}

	 \item "data-classic-seed-lots"

	 \item "data-classic-relation"

	 \item "data-S-seed-lots"

	 \item "data-S-relation"

	 \item "data-SR-seed-lots"

	 \item "data-SR-relation"

	}
}
\examples{
# See the vignette
}
\author{
Pierre Riviere
}
\seealso{
\code{\link{encrypt.data}}, \code{\link{translate.data}},, \code{\link{format.data}}, \code{\link{get.ggplot}}, \code{\link{get.table}}
}

